refactor: split compaction.test.ts into dedicated test files
