+++ packages/agent-core/test/agent/compaction/full.test.ts
@@ -12,12 +12,12 @@ import {
 } from '@moonshot-ai/kosong';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import type { AgentOptions } from '../../src/agent';
-import { DefaultCompactionStrategy, type CompactionStrategy } from '../../src/agent/compaction';
-import { HookEngine, type HookEngineTriggerArgs } from '../../src/session/hooks';
-import { recordingTelemetry, type TelemetryRecord } from '../fixtures/telemetry';
-import type { TestAgentContext, TestAgentOptions } from './harness/agent';
-import { testAgent } from './harness/agent';
+import type { AgentOptions } from '../../../src/agent';
+import { DefaultCompactionStrategy, type CompactionStrategy } from '../../../src/agent/compaction';
+import { HookEngine, type HookEngineTriggerArgs } from '../../../src/session/hooks';
+import { recordingTelemetry, type TelemetryRecord } from '../../fixtures/telemetry';
+import type { TestAgentContext, TestAgentOptions } from '../harness/agent';
+import { testAgent } from '../harness/agent';
 
 type GenerateFn = NonNullable<AgentOptions['generate']>;
 
@@ -35,120 +35,7 @@ const CATALOGUED_MODEL_CAPABILITIES = {
   max_context_tokens: 256_000,
 } as const;
 
-describe('Agent compaction', () => {
-  it('keeps an oversized trailing user message as recent', () => {
-    const strategy = testCompactionStrategy();
-    const messages = [
-      textMessage('user', 'old user'),
-      textMessage('assistant', 'old assistant'),
-      textMessage('user', `pending user ${'x'.repeat(1_200)}`),
-    ];
-
-    expect(strategy.computeCompactCount(messages, 'auto')).toBe(2);
-  });
-
-  it('keeps consecutive trailing user messages as recent', () => {
-    const strategy = testCompactionStrategy();
-    const messages = [
-      textMessage('user', 'old user'),
-      textMessage('assistant', 'old assistant'),
-      textMessage('user', `pending user one ${'x'.repeat(1_200)}`),
-      textMessage('user', `pending user two ${'x'.repeat(1_200)}`),
-    ];
-
-    expect(strategy.computeCompactCount(messages, 'auto')).toBe(2);
-  });
-
-  it('compacts the prefix when the trailing exchange itself is oversized', () => {
-    const strategy = testCompactionStrategy();
-    const messages = [
-      textMessage('user', 'old user'),
-      textMessage('assistant', 'old assistant'),
-      textMessage('user', 'recent user'),
-      textMessage('assistant', `recent assistant ${'x'.repeat(1_200)}`),
-    ];
-
-    expect(strategy.computeCompactCount(messages, 'auto')).toBe(2);
-  });
-
-  it('returns 0 when there is nothing to compact', () => {
-    const strategy = testCompactionStrategy();
-    expect(strategy.computeCompactCount([], 'auto')).toBe(0);
-    expect(strategy.computeCompactCount([textMessage('user', 'only pending')], 'auto')).toBe(0);
-    expect(
-      strategy.computeCompactCount(
-        [
-          textMessage('user', 'a'),
-          textMessage('user', 'b'),
-          textMessage('user', 'c'),
-        ],
-        'auto',
-      ),
-    ).toBe(0);
-  });
-
-  it('returns 0 when no intermediate split exists and the last message is also unsplittable', () => {
-    const strategy = testCompactionStrategy();
-    const messages: Message[] = [
-      textMessage('user', 'inspect'),
-      {
-        role: 'assistant',
-        content: [],
-        toolCalls: [{ type: 'function', id: 'call_a', name: 'Lookup', arguments: '{}' }],
-      },
-    ];
-
-    expect(strategy.computeCompactCount(messages, 'auto')).toBe(0);
-  });
-
-  it('does not split inside a parallel tool exchange', () => {
-    const strategy = testCompactionStrategy();
-    const messages: Message[] = [
-      textMessage('user', 'old user'),
-      textMessage('assistant', 'old assistant'),
-      textMessage('user', 'run both tools'),
-      {
-        role: 'assistant',
-        content: [],
-        toolCalls: [
-          { type: 'function', id: 'call_a', name: 'Lookup', arguments: '{}' },
-          { type: 'function', id: 'call_b', name: 'Lookup', arguments: '{}' },
-        ],
-      },
-      { role: 'tool', content: [{ type: 'text', text: 'a' }], toolCalls: [], toolCallId: 'call_a' },
-      { role: 'tool', content: [{ type: 'text', text: 'b' }], toolCalls: [], toolCallId: 'call_b' },
-      textMessage('user', 'next prompt'),
-    ];
-
-    // The only valid split is before the parallel exchange (after 'old assistant'),
-    // never between tool_a and tool_b — that would leave tool_b as an orphan.
-    expect(strategy.computeCompactCount(messages, 'auto')).toBe(2);
-  });
-
-  it('reserves response context by default before the ratio threshold is reached', () => {
-    const strategy = new DefaultCompactionStrategy(() => 256_000);
-
-    expect(strategy.shouldCompact(210_000)).toBe(true);
-    expect(strategy.shouldBlock(210_000)).toBe(true);
-  });
-
-  it('ignores reserved context when the reserve is not smaller than the model window', () => {
-    const strategy = new DefaultCompactionStrategy(() => 32_000, {
-      triggerRatio: 0.85,
-      blockRatio: 0.85,
-      reservedContextSize: 50_000,
-      maxCompactionPerTurn: 3,
-      maxRecentMessages: 3,
-      maxRecentUserMessages: Infinity,
-      maxRecentSizeRatio: 0.2,
-    });
-
-    expect(strategy.shouldCompact(1)).toBe(false);
-    expect(strategy.shouldBlock(1)).toBe(false);
-    expect(strategy.shouldCompact(28_000)).toBe(true);
-    expect(strategy.shouldBlock(28_000)).toBe(true);
-  });
-
+describe('FullCompaction', () => {
   it('runs manual compaction and applies the compacted context', async () => {
     const records: TelemetryRecord[] = [];
     const ctx = testAgent({ telemetry: recordingTelemetry(records) });
@@ -1620,219 +1507,3 @@ function inputHistorySnapshot(history: readonly Message[]): string[] {
 function normalizeInputText(text: string): string {
   return text.includes('compact this conversation context') ? '<compaction-instruction>' : text;
 }
-
-describe('MicroCompaction', () => {
-  it('truncates old tool results after cache miss', () => {
-    vi.useFakeTimers();
-    const ctx = testAgent({
-      microCompaction: {
-        keepRecentMessages: 4,
-        minContentTokens: 1,
-        cacheMissedThresholdMs: 60 * 60 * 1000,
-      },
-    });
-
-    vi.setSystemTime(0);
-    ctx.appendToolExchange();
-    ctx.appendToolExchange();
-    ctx.appendToolExchange();
-
-    expect(ctx.agent.context.messages).toHaveLength(9);
-
-    vi.setSystemTime(61 * 60 * 1000);
-
-    const messages = ctx.agent.context.messages;
-    const marker = '[Old tool result content cleared]';
-
-    expect(messages[2]).toMatchObject({
-      role: 'tool',
-      content: [{ type: 'text', text: marker }],
-    });
-    expect(messages[5]).toMatchObject({
-      role: 'tool',
-      content: [{ type: 'text', text: 'lookup result' }],
-    });
-    expect(messages[8]).toMatchObject({
-      role: 'tool',
-      content: [{ type: 'text', text: 'lookup result' }],
-    });
-  });
-
-  it('does nothing before cache miss threshold', () => {
-    vi.useFakeTimers();
-    const ctx = testAgent({
-      microCompaction: {
-        keepRecentMessages: 4,
-        minContentTokens: 1,
-        cacheMissedThresholdMs: 60 * 60 * 1000,
-      },
-    });
-
-    vi.setSystemTime(0);
-    ctx.appendToolExchange();
-    ctx.appendToolExchange();
-    ctx.appendToolExchange();
-
-    vi.setSystemTime(30 * 60 * 1000);
-
-    const messages = ctx.agent.context.messages;
-    expect(messages.every((m) => m.role !== 'tool' || (m.content[0] as { text: string })?.text !== '[Old tool result content cleared]')).toBe(true);
-  });
-
-  it('persists cutoff across calls until cache miss resets it', () => {
-    vi.useFakeTimers();
-    const ctx = testAgent({
-      microCompaction: {
-        keepRecentMessages: 2,
-        minContentTokens: 1,
-        cacheMissedThresholdMs: 60 * 60 * 1000,
-      },
-    });
-
-    vi.setSystemTime(0);
-    ctx.appendToolExchange();
-    ctx.appendToolExchange();
-
-    vi.setSystemTime(61 * 60 * 1000);
-
-    const first = ctx.agent.context.messages;
-    expect(first[2]).toMatchObject({
-      role: 'tool',
-      content: [{ type: 'text', text: '[Old tool result content cleared]' }],
-    });
-
-    vi.setSystemTime(62 * 60 * 1000);
-
-    const second = ctx.agent.context.messages;
-    expect(second[2]).toMatchObject({
-      role: 'tool',
-      content: [{ type: 'text', text: '[Old tool result content cleared]' }],
-    });
-  });
-
-  it('clears cutoff on reset', () => {
-    vi.useFakeTimers();
-    const ctx = testAgent({
-      microCompaction: {
-        keepRecentMessages: 4,
-        minContentTokens: 1,
-        cacheMissedThresholdMs: 60 * 60 * 1000,
-      },
-    });
-
-    vi.setSystemTime(0);
-    ctx.appendToolExchange();
-    ctx.appendToolExchange();
-
-    vi.setSystemTime(61 * 60 * 1000);
-
-    ctx.agent.microCompaction.reset();
-
-    const messages = ctx.agent.context.messages;
-    expect(messages.every((m) => m.role !== 'tool' || (m.content[0] as { text: string })?.text !== '[Old tool result content cleared]')).toBe(true);
-  });
-
-  it('skips tool results below minContentTokens', () => {
-    vi.useFakeTimers();
-    const ctx = testAgent({
-      microCompaction: {
-        keepRecentMessages: 2,
-        minContentTokens: 100,
-        cacheMissedThresholdMs: 60 * 60 * 1000,
-      },
-    });
-
-    vi.setSystemTime(0);
-    ctx.appendToolExchange();
-    ctx.appendToolExchange();
-
-    vi.setSystemTime(61 * 60 * 1000);
-
-    const messages = ctx.agent.context.messages;
-    expect(
-      messages.every(
-        (m) => m.role !== 'tool' || (m.content[0] as { text: string })?.text !== '[Old tool result content cleared]',
-      ),
-    ).toBe(true);
-  });
-
-  it('skips non-tool messages', () => {
-    vi.useFakeTimers();
-    const ctx = testAgent({
-      microCompaction: {
-        keepRecentMessages: 2,
-        minContentTokens: 1,
-        cacheMissedThresholdMs: 60 * 60 * 1000,
-      },
-    });
-
-    vi.setSystemTime(0);
-    ctx.appendExchange(1, 'user one', 'assistant one', 10);
-    ctx.appendExchange(2, 'user two', 'assistant two', 10);
-    ctx.appendExchange(3, 'user three', 'assistant three', 10);
-
-    vi.setSystemTime(61 * 60 * 1000);
-
-    const messages = ctx.agent.context.messages;
-    expect(messages.every((m) => m.role === 'user' || m.role === 'assistant')).toBe(true);
-    expect(
-      messages.every(
-        (m) => m.role !== 'tool' || (m.content[0] as { text: string })?.text !== '[Old tool result content cleared]',
-      ),
-    ).toBe(true);
-  });
-
-  it('clears cutoff on context clear', () => {
-    vi.useFakeTimers();
-    const ctx = testAgent({
-      microCompaction: {
-        keepRecentMessages: 2,
-        minContentTokens: 1,
-        cacheMissedThresholdMs: 60 * 60 * 1000,
-      },
-    });
-
-    vi.setSystemTime(0);
-    ctx.appendToolExchange();
-    ctx.appendToolExchange();
-
-    vi.setSystemTime(61 * 60 * 1000);
-
-    ctx.agent.context.clear();
-
-    expect(ctx.agent.context.messages).toHaveLength(0);
-    expect(ctx.agent.context.lastAssistantAt).toBe(0);
-  });
-
-  it('clears cutoff on full compaction', async () => {
-    vi.useFakeTimers();
-    const ctx = testAgent({
-      microCompaction: {
-        keepRecentMessages: 2,
-        minContentTokens: 1,
-        cacheMissedThresholdMs: 60 * 60 * 1000,
-      },
-    });
-    ctx.configure({
-      provider: CATALOGUED_PROVIDER,
-      modelCapabilities: CATALOGUED_MODEL_CAPABILITIES,
-    });
-
-    vi.setSystemTime(0);
-    ctx.appendExchange(1, 'old user', 'old assistant', 20);
-    ctx.appendExchange(2, 'recent user', 'recent assistant', 80);
-
-    vi.setSystemTime(61 * 60 * 1000);
-
-    const compacted = ctx.once('context.apply_compaction');
-    ctx.mockNextResponse({ type: 'text', text: 'Summary.' });
-    await ctx.rpc.beginCompaction({});
-    await compacted;
-
-    expect(ctx.agent.context.messages).toHaveLength(1);
-    expect(ctx.agent.context.messages[0]).toMatchObject({
-      role: 'assistant',
-      content: [{ type: 'text', text: 'Summary.' }],
-    });
-  });
-});
