+++ packages/agent-core/test/agent/compaction/strategy.test.ts
@@ -1,0 +1,183 @@
+import { existsSync, mkdtempSync, readFileSync } from 'node:fs';
+import { tmpdir } from 'node:os';
+import { join } from 'pathe';
+
+import {
+  APIConnectionError,
+  APIContextOverflowError,
+  APIStatusError,
+  UNKNOWN_CAPABILITY,
+  type Message,
+  type ToolCall,
+} from '@moonshot-ai/kosong';
+import { afterEach, describe, expect, it, vi } from 'vitest';
+
+import type { AgentOptions } from '../../../src/agent';
+import { DefaultCompactionStrategy, type CompactionStrategy } from '../../../src/agent/compaction';
+import { HookEngine, type HookEngineTriggerArgs } from '../../../src/session/hooks';
+import { recordingTelemetry, type TelemetryRecord } from '../../fixtures/telemetry';
+import type { TestAgentContext, TestAgentOptions } from '../harness/agent';
+import { testAgent } from '../harness/agent';
+
+type GenerateFn = NonNullable<AgentOptions['generate']>;
+
+const CATALOGUED_PROVIDER = {
+  type: 'kimi',
+  apiKey: 'test-key',
+  model: 'kimi-code',
+} as const;
+const CATALOGUED_MODEL_CAPABILITIES = {
+  image_in: true,
+  video_in: true,
+  audio_in: false,
+  thinking: true,
+  tool_use: true,
+  max_context_tokens: 256_000,
+} as const;
+
+describe('DefaultCompactionStrategy', () => {
+  it('keeps an oversized trailing user message as recent', () => {
+    const strategy = testCompactionStrategy();
+    const messages = [
+      textMessage('user', 'old user'),
+      textMessage('assistant', 'old assistant'),
+      textMessage('user', `pending user ${'x'.repeat(1_200)}`),
+    ];
+
+    expect(strategy.computeCompactCount(messages, 'auto')).toBe(2);
+  });
+
+  it('keeps consecutive trailing user messages as recent', () => {
+    const strategy = testCompactionStrategy();
+    const messages = [
+      textMessage('user', 'old user'),
+      textMessage('assistant', 'old assistant'),
+      textMessage('user', `pending user one ${'x'.repeat(1_200)}`),
+      textMessage('user', `pending user two ${'x'.repeat(1_200)}`),
+    ];
+
+    expect(strategy.computeCompactCount(messages, 'auto')).toBe(2);
+  });
+
+  it('compacts the prefix when the trailing exchange itself is oversized', () => {
+    const strategy = testCompactionStrategy();
+    const messages = [
+      textMessage('user', 'old user'),
+      textMessage('assistant', 'old assistant'),
+      textMessage('user', 'recent user'),
+      textMessage('assistant', `recent assistant ${'x'.repeat(1_200)}`),
+    ];
+
+    expect(strategy.computeCompactCount(messages, 'auto')).toBe(2);
+  });
+
+  it('returns 0 when there is nothing to compact', () => {
+    const strategy = testCompactionStrategy();
+    expect(strategy.computeCompactCount([], 'auto')).toBe(0);
+    expect(strategy.computeCompactCount([textMessage('user', 'only pending')], 'auto')).toBe(0);
+    expect(
+      strategy.computeCompactCount(
+        [
+          textMessage('user', 'a'),
+          textMessage('user', 'b'),
+          textMessage('user', 'c'),
+        ],
+        'auto',
+      ),
+    ).toBe(0);
+  });
+
+  it('returns 0 when no intermediate split exists and the last message is also unsplittable', () => {
+    const strategy = testCompactionStrategy();
+    const messages: Message[] = [
+      textMessage('user', 'inspect'),
+      {
+        role: 'assistant',
+        content: [],
+        toolCalls: [{ type: 'function', id: 'call_a', name: 'Lookup', arguments: '{}' }],
+      },
+    ];
+
+    expect(strategy.computeCompactCount(messages, 'auto')).toBe(0);
+  });
+
+  it('does not split inside a parallel tool exchange', () => {
+    const strategy = testCompactionStrategy();
+    const messages: Message[] = [
+      textMessage('user', 'old user'),
+      textMessage('assistant', 'old assistant'),
+      textMessage('user', 'run both tools'),
+      {
+        role: 'assistant',
+        content: [],
+        toolCalls: [
+          { type: 'function', id: 'call_a', name: 'Lookup', arguments: '{}' },
+          { type: 'function', id: 'call_b', name: 'Lookup', arguments: '{}' },
+        ],
+      },
+      { role: 'tool', content: [{ type: 'text', text: 'a' }], toolCalls: [], toolCallId: 'call_a' },
+      { role: 'tool', content: [{ type: 'text', text: 'b' }], toolCalls: [], toolCallId: 'call_b' },
+      textMessage('user', 'next prompt'),
+    ];
+
+    // The only valid split is before the parallel exchange (after 'old assistant'),
+    // never between tool_a and tool_b — that would leave tool_b as an orphan.
+    expect(strategy.computeCompactCount(messages, 'auto')).toBe(2);
+  });
+
+  it('reserves response context by default before the ratio threshold is reached', () => {
+    const strategy = new DefaultCompactionStrategy(() => 256_000);
+
+    expect(strategy.shouldCompact(210_000)).toBe(true);
+    expect(strategy.shouldBlock(210_000)).toBe(true);
+  });
+
+  it('ignores reserved context when the reserve is not smaller than the model window', () => {
+    const strategy = new DefaultCompactionStrategy(() => 32_000, {
+      triggerRatio: 0.85,
+      blockRatio: 0.85,
+      reservedContextSize: 50_000,
+      maxCompactionPerTurn: 3,
+      maxRecentMessages: 3,
+      maxRecentUserMessages: Infinity,
+      maxRecentSizeRatio: 0.2,
+    });
+
+    expect(strategy.shouldCompact(1)).toBe(false);
+    expect(strategy.shouldBlock(1)).toBe(false);
+    expect(strategy.shouldCompact(28_000)).toBe(true);
+    expect(strategy.shouldBlock(28_000)).toBe(true);
+  });
+});
+
+function testCompactionStrategy(maxSize: number = 1_000): DefaultCompactionStrategy {
+  return new DefaultCompactionStrategy(() => maxSize, {
+    triggerRatio: 0.85,
+    blockRatio: 0.85,
+    reservedContextSize: 0,
+    maxCompactionPerTurn: 3,
+    maxRecentMessages: 10,
+    maxRecentUserMessages: Infinity,
+    maxRecentSizeRatio: 0.2,
+  });
+}
+
+function overflowOnlyCompactionStrategy(maxSize: number = 14): DefaultCompactionStrategy {
+  return new DefaultCompactionStrategy(() => maxSize, {
+    triggerRatio: Infinity,
+    blockRatio: Infinity,
+    reservedContextSize: 0,
+    maxCompactionPerTurn: 3,
+    maxRecentMessages: 3,
+    maxRecentUserMessages: Infinity,
+    maxRecentSizeRatio: 0.2,
+  });
+}
+
+function textMessage(role: 'user' | 'assistant', text: string): Message {
+  return {
+    role,
+    content: [{ type: 'text', text }],
+    toolCalls: [],
+  };
+}
